Finished the implementation! Round I
This commit fixes #17, #19, #20.

Now butterfly can generate the whole project from the scratch.
createSimpleBoxCase.gh is a prototype in Grasshopper that shows how this
can all go together.
Haven't still tried to run it. There are probably bugs in the files but
the main part of the library is finished.