Added refinement regions to butterfly
Both Case from surfaces and wind tunnel component accept refinment
regions as a list.

There are three new components that let you create a refinement region
from any arbitery shape. mode can be inside/outside or distance.

Geometry library is now better structured as BFMesh,
RefinementRegion<BFMesh>, BFGeometry<BFMesh>,
BFBlockGeometry<BFGeometry>

There is a similar approach on Grasshopper side except that GHBFGeometry
also is a suclass of BFMesh and not GHMesh. There is a class for GHMesh
which is used to analyze input geometries and create a joined mesh that
can be converted to an OpenFOAM mesh.

Resolved #59.