Finished 1st round of implementation for parallel
This commit addresses #60. The parallel run has been implemented both
for snappyHexMesh and solution. I personally feel the current
implementation is not flexible enough. Each time that the user stops the
solution butterfly put the folders back together, and decompose the mesh
again for continuing the analysis. This is niether smart nor efficient.
I'm thinking to take out reconstructPar and remover processes to a new
component for a better control.