Started adding heatTransfer recipe
This commit starts to address #86. Before pushing this to master I need
to address #84. The back and forth to set up fvSchemes and fvSolution is
too much with the current implementation and can genrate invalid cases.
I will keep this local until I address the issue that are mentioned
above.