Fixed bug with convertToMeters
Resolved #94
OpenFOAM is looking for x, y, z values in meters even if the
converToMeters is not 1. This is now added to the code to make sure
locationInMesh will be converted correctly.
Also added scaling to loadMesh and loadPoints so they will be scaled to
Rhino units from butterfly case.